Update inference checks to work with new cpp specification
