+++ R_ignore/inference_checks/basic_model.Rmd
@@ -29,13 +29,7 @@ data_list <- list(x = rnorm(n))
 Likelihood and prior:
 
 ```{r, echo = FALSE, comment = ''}
-likelihood_filename <- "source/basic_loglike.cpp"
-cpp_loglike <- cpp_function_get(likelihood_filename)
-cat(readLines(likelihood_filename), sep = '\n')
-
-prior_filename <- "source/basic_logprior.cpp"
-cpp_logprior <- cpp_function_get(prior_filename)
-cat(readLines(prior_filename), sep = '\n')
+Rcpp::sourceCpp("source/basic_loglike_logprior.cpp")
 ```
 
 Parameters dataframe:
@@ -49,8 +43,8 @@ df_params <- define_params(name = "mu", min = -1, max = 1)
 ```{r cars}
 mcmc <- run_mcmc(data = data_list,
                  df_params = df_params,
-                 loglike = cpp_loglike,
-                 logprior = cpp_logprior,
+                 loglike = "loglike",
+                 logprior = "logprior",
                  burnin = 1e3,
                  samples = 1e5,
                  chains = 10,
