+++ R_ignore/inference_checks/return_prior.Rmd
@@ -32,13 +32,7 @@ df_params <- define_params(name = "real_line", min = -Inf, max = Inf,
 Likelihood and prior:
 
 ```{r, echo = FALSE, comment = ''}
-likelihood_filename <- "source/returnprior_loglike.cpp"
-cpp_loglike <- cpp_function_get(likelihood_filename)
-cat(readLines(likelihood_filename), sep = '\n')
-
-prior_filename <- "source/returnprior_logprior.cpp"
-cpp_logprior <- cpp_function_get(prior_filename)
-cat(readLines(prior_filename), sep = '\n')
+Rcpp::sourceCpp("source/returnprior_loglike_logprior.cpp")
 ```
 
 ## Run MCMC
@@ -46,8 +40,8 @@ cat(readLines(prior_filename), sep = '\n')
 ```{r}
 mcmc <- run_mcmc(data = list(x = -1),
                  df_params = df_params,
-                 loglike = cpp_loglike,
-                 logprior = cpp_logprior,
+                 loglike = "loglike",
+                 logprior = "logprior",
                  burnin = 1e3,
                  samples = 1e5,
                  chains = 10,
