+++ R_ignore/inference_checks/source/basic_loglike_logprior.cpp
@@ -1,0 +1,40 @@
+#include <Rcpp.h>
+using namespace Rcpp;
+
+// [[Rcpp::export]]
+SEXP loglike(Rcpp::NumericVector params, Rcpp::List data, Rcpp::List misc) {
+  
+  std::vector<double> x = Rcpp::as< std::vector<double> >(data["x"]);
+  
+  double ret = 0.0;
+  for (unsigned int i = 0; i < x.size(); ++i) {
+    ret += R::dnorm(x[i], params["mu"], 1.0, true);
+  }
+  
+  return Rcpp::wrap(ret);
+}
+
+// [[Rcpp::export]]
+SEXP logprior(Rcpp::NumericVector params, Rcpp::List misc) {
+  return Rcpp::wrap(0.0);
+}
+
+// [[Rcpp::export]]  
+SEXP create_xptr_loglike(std::string fstr) {  
+  typedef SEXP (*funcPtr)(Rcpp::NumericVector params, Rcpp::List data, Rcpp::List misc) ;  
+  if (fstr == "loglike"){
+    return(Rcpp::XPtr<funcPtr>(new funcPtr(&loglike))) ;
+  } else {
+    return(Rcpp::XPtr<funcPtr>(R_NilValue)) ; 
+  }
+}
+
+// [[Rcpp::export]]  
+SEXP create_xptr_logprior(std::string fstr) {  
+  typedef SEXP (*funcPtr)(Rcpp::NumericVector params, Rcpp::List misc) ;  
+  if (fstr == "logprior"){
+    return(Rcpp::XPtr<funcPtr>(new funcPtr(&logprior))) ;
+  } else {
+    return(Rcpp::XPtr<funcPtr>(R_NilValue)) ; 
+  }
+}
