--- R_ignore/inference_checks/source/multilevel_logprior.cpp
@@ -1,9 +1,0 @@
-#include <Rcpp.h>
-using namespace Rcpp;
-
-// [[Rcpp::export]]
-// drjacoby_function_start
-SEXP logprior(Rcpp::NumericVector params, Rcpp::List misc) {
-  return Rcpp::wrap(0.0);
-}
-// drjacoby_function_end
