--- R_ignore/inference_checks/source/returnprior_loglike.cpp
@@ -1,9 +1,0 @@
-#include <Rcpp.h>
-using namespace Rcpp;
-
-// [[Rcpp::export]]
-// drjacoby_function_start
-SEXP loglike(Rcpp::NumericVector params, Rcpp::List data, Rcpp::List misc) {
-  return Rcpp::wrap(0.0);
-}
-// drjacoby_function_end
