+++ R_ignore/inference_checks/source/returnprior_loglike_logprior.cpp
@@ -1,0 +1,38 @@
+#include <Rcpp.h>
+using namespace Rcpp;
+
+// [[Rcpp::export]]
+SEXP loglike(Rcpp::NumericVector params, Rcpp::List data, Rcpp::List misc) {
+  return Rcpp::wrap(0.0);
+}
+
+// [[Rcpp::export]]
+SEXP logprior(Rcpp::NumericVector params, Rcpp::List misc) {
+  
+  double ret = R::dnorm(params["real_line"], 0.0, 1.0, true) +
+    R::dgamma(-params["neg_line"], 5.0, 5.0, true) +
+    R::dgamma(params["pos_line"], 5.0, 5.0, true) +
+    R::dbeta(params["unit_interval"], 3.0, 3.0, true);
+  
+  return Rcpp::wrap(ret);
+}
+
+// [[Rcpp::export]]  
+SEXP create_xptr_loglike(std::string fstr) {  
+  typedef SEXP (*funcPtr)(Rcpp::NumericVector params, Rcpp::List data, Rcpp::List misc) ;  
+  if (fstr == "loglike"){
+    return(Rcpp::XPtr<funcPtr>(new funcPtr(&loglike))) ;
+  } else {
+    return(Rcpp::XPtr<funcPtr>(R_NilValue)) ; 
+  }
+}
+
+// [[Rcpp::export]]  
+SEXP create_xptr_logprior(std::string fstr) {  
+  typedef SEXP (*funcPtr)(Rcpp::NumericVector params, Rcpp::List misc) ;  
+  if (fstr == "logprior"){
+    return(Rcpp::XPtr<funcPtr>(new funcPtr(&logprior))) ;
+  } else {
+    return(Rcpp::XPtr<funcPtr>(R_NilValue)) ; 
+  }
+}
